Update to latest Transit

+++ Test/OtherModeBuilderTest.cs
@@ -20,11 +20,11 @@ namespace Test
             
             var cf = crow as CrowsFlightTransferGenerator;
             Assert.NotNull(cf);
-            Assert.Equal(500, cf.Range());
+            Assert.Equal((uint)500, cf.Range());
             var crow0 = omb.Create("crowsflight&speed=1&maxDistance=1500", empty, empty);
             var cf0 = crow0 as CrowsFlightTransferGenerator;
             Assert.NotNull(cf0);
-            Assert.Equal(1500, cf0.Range());
+            Assert.Equal((uint) 1500, cf0.Range());
         }
 
         [Fact]
@@ -44,7 +44,7 @@ namespace Test
             var gen = omb.Create(desc, empty, empty);
             var flm = gen as FirstLastMilePolicy;
             Assert.NotNull(flm);
-            Assert.Equal(5000, flm.Range());
+            Assert.Equal((uint) 5000, flm.Range());
 
 
             var exp = "firstLastMile" +
